Port gpc test to Playwright
