+++ integration-test/data/pages/blank_framer.html
